rename "flags" module to "parse_flags"
